Add auth, fix stuff

+++ lib/veranxiety/allergy.ex
@@ -5,88 +5,103 @@ defmodule Veranxiety.Allergy do
 
   import Ecto.Query, warn: false
   alias Veranxiety.Repo
-
   alias Veranxiety.Allergy.Entry
+  # Assuming User module exists
+  alias Veranxiety.Accounts.User
 
   @doc """
-  Returns the list of allergy_entries.
+  Returns the list of allergy entries for a specific user.
 
   ## Examples
 
-      iex> list_allergy_entries()
+      iex> list_allergy_entries(user)
       [%Entry{}, ...]
 
   """
-  def list_allergy_entries do
-    Repo.all(Entry)
+  def list_allergy_entries(%User{} = user) do
+    Entry
+    |> where(user_id: ^user.id)
+    |> Repo.all()
   end
 
   @doc """
-  Gets a single entry.
+  Gets a single entry for a specific user.
 
-  Raises `Ecto.NoResultsError` if the Entry does not exist.
+  Raises `Ecto.NoResultsError` if the Entry does not exist or does not belong to the user.
 
   ## Examples
 
-      iex> get_entry!(123)
+      iex> get_entry!(user, 123)
       %Entry{}
 
-      iex> get_entry!(456)
+      iex> get_entry!(user, 456)
       ** (Ecto.NoResultsError)
 
   """
-  def get_entry!(id), do: Repo.get!(Entry, id)
+  def get_entry!(%User{} = user, id) do
+    Entry
+    |> where([e], e.id == ^id and e.user_id == ^user.id)
+    |> Repo.one!()
+  end
 
   @doc """
-  Creates a entry.
+  Creates an entry associated with a user.
 
   ## Examples
 
-      iex> create_entry(%{field: value})
+      iex> create_entry(user, %{field: value})
       {:ok, %Entry{}}
 
-      iex> create_entry(%{field: bad_value})
+      iex> create_entry(user, %{field: bad_value})
       {:error, %Ecto.Changeset{}}
 
   """
-  def create_entry(attrs \\ %{}) do
-    %Entry{}
+  def create_entry(%User{} = user, attrs \\ %{}) do
+    %Entry{user_id: user.id}
     |> Entry.changeset(attrs)
     |> Repo.insert()
   end
 
   @doc """
-  Updates a entry.
+  Updates an entry for a specific user.
 
   ## Examples
 
-      iex> update_entry(entry, %{field: new_value})
+      iex> update_entry(user, entry, %{field: new_value})
       {:ok, %Entry{}}
 
-      iex> update_entry(entry, %{field: bad_value})
+      iex> update_entry(user, entry, %{field: bad_value})
       {:error, %Ecto.Changeset{}}
 
   """
-  def update_entry(%Entry{} = entry, attrs) do
-    entry
-    |> Entry.changeset(attrs)
-    |> Repo.update()
+  def update_entry(%User{} = user, %Entry{} = entry, attrs) do
+    if entry.user_id == user.id do
+      entry
+      |> Entry.changeset(attrs)
+      |> Repo.update()
+    else
+      {:error, :unauthorized}
+    end
   end
 
   @doc """
-  Deletes a entry.
+  Deletes an entry for a specific user.
 
   ## Examples
 
-      iex> delete_entry(entry)
+      iex> delete_entry(user, entry)
       {:ok, %Entry{}}
 
-      iex> delete_entry(entry)
+      iex> delete_entry(user, entry)
       {:error, %Ecto.Changeset{}}
 
   """
-  def delete_entry(%Entry{} = entry) do
-    Repo.delete(entry)
+  def delete_entry(%User{} = user, %Entry{} = entry) do
+    if entry.user_id == user.id do
+      Repo.delete(entry)
+    else
+      {:error, :unauthorized}
+    end
   end
 
   @doc """
